the_data: main keystroke timing data

This data frame contains the pre-processed data used for the bulk of the empirical analyses in Crump, Lai, & Brosowsky. The data was originally collected by Behmer & Crump (2017).

This data frame contains individual keystroke typing data from 346 typists who copy typed 5 English paragraphs. It has been filtered to remove keystrokes from capitalized letters.

the_data

Format

A data frame with 1017214 rows and 27 variables:

Subject

integer A unique number coding each subject

ParagraphType

character "N" stands for normal english. This data frame only contains data from yping English paragraphs. The raw data contains additional typing data from typing random letter strings, and english-like strings.

IKSIs

integer Interkeystroke interval in milliseconds, the time elapsed between the previous keystroke and the current keystroke.

Letters

character the character typed

PredBigram

character bigram showing the previous character and current character

SuccBigram

character bigram showing the current character and the next character

PredBigramCorrect

character Accuracy codes for the previous letter and current letter, 1 = correct, 0 = typo

SuccBigramCorrect

character Accuracy codes for the current letter and next letter, 1 = correct, 0 = typo

PredTrigram

character trigram for the previous two letters and current letter

SuccTrigram

character trigram for the current letter and next two letter

PredTrigramCorrect

character accuracy codes for previous two letters and current letter, 1 = correct, 0 = typo

SuccTrigramCorrect

character accuracy codes for current letter and next two letters, 1 = correct, 0 = typo

GUT

integer Letter frequencies estimated from the gutenberg corpus

PredBigramMLE

double Probability of current letter given previous letter, from gutenberg corpus

SuccBigramMLE

double Probability of current letter given next letter, from gutenberg corpus

PredBigramFreq

integer preceding bigram frequency from gutenberg

SuccBigramFreq

integer succeeding bigram frequency from gutenberg

PredTrigramMLE

double probability of current letter given preceding trigram

PredTrigramFreq

integer preceding trigram frequency from gutenberg

SuccTrigramMLE

double probability of current letter given succeeding trigram

SuccTrigramFreq

integer succeeding trigram frequency from gutenberg

let_pos

integer serial position of letter within word

word_lengths

integer number of letters in word

probs

double need to double-check what this is

whole_word

character the current word, note any word that begins with a capital letter is displayed with the capital letter removed

word_prob

double need to double-check what this is.

whole_word_acc

character 1 = the whole word was typed correctly, 0 = there was an error(s) somewhere in the word